Now using an XML parser to create a scene from an XML-file.

+++ external_libraries/common_include/objloader.h
@@ -17,7 +17,7 @@ bool loadOBJ(
 
 bool loadAssImp(
 	const char * path, 
-	std::vector<unsigned short> & indices,
+	std::vector<unsigned int> & indices,
 	std::vector<glm::vec3> & vertices,
 	std::vector<glm::vec2> & uvs,
 	std::vector<glm::vec3> & normals
